Again DO NOT PR THIS BRANCH
Just committing so that I have a backup as I start messing with some other things - this is still NOT ready for merging so needs to remain quarantined.

For reference: This includes some modular implementations to extend the fitting procedures to blocks beyond LinearNeuralMass. As it is, there's hard-coded state number issues with the old implementation, and hierarchy issues with the `system_from_graph` implementation. So we'll need to talk through some design choices after the deadline is met for other stuff.