Make Izhikevich neurons consistent with CC paper
The default neurons are slightly different than the Chen/Campbell neurons to be consistent with each other, but we'll need these to run larger simulations.

This commit makes the neurons for a standard implementation, and they look like they're operating as expected (bursting together) for a small population. Need GraphDynamics for larger one so working on that next.